@@ -5,126 +5,60 @@ import os
 import sys
 import mmap
 import codecs
-import string
 import logging
 import argparse
 import textwrap
 import contextlib
 from enum import Enum
 from time import time
-from typing import Set, List, Union, Optional
+from typing import Set, List, Optional
 
 import halo
-import tqdm
-import tabulate
 import viv_utils
 import viv_utils.flirt
 from vivisect import VivWorkspace
 
-import floss.logging
-import floss.strings as strings
+import floss.utils
+import floss.results
 import floss.version
+import floss.logging_
 import floss.render.json
-import floss.stackstrings as stackstrings
-import floss.string_decoder as string_decoder
-from floss.const import MAX_FILE_SIZE, DEFAULT_MIN_LENGTH, SUPPORTED_FILE_MAGIC
-from floss.utils import hex, get_vivisect_meta_info
-from floss.results import Metadata, AddressType, StackString, TightString, DecodedString, ResultDocument, StringEncoding
+import floss.render.default
+from floss.const import MAX_FILE_SIZE, MIN_STRING_LENGTH, SUPPORTED_FILE_MAGIC
+from floss.utils import hex, get_runtime_diff, get_vivisect_meta_info, set_vivisect_log_level
+from floss.results import Metadata, ResultDocument
+from floss.strings import extract_ascii_unicode_strings
 from floss.version import __version__
 from floss.identify import (
+    append_unique,
+    get_function_fvas,
     get_top_functions,
+    get_tight_function_fvas,
     get_functions_with_tightloops,
     find_decoding_function_features,
     get_functions_without_tightloops,
 )
+from floss.logging_ import DebugLevel
+from floss.stackstrings import extract_stackstrings
 from floss.tightstrings import extract_tightstrings
-
-DEFAULT_MAX_INSN_COUNT = 20000
-DEFAULT_MAX_ADDRESS_REVISITS = 0
-
-DEFAULT_SHELLCODE_ARCH = "auto"
-DEFAULT_SHELLCODE_BASE = 0x1000
-DEFAULT_SHELLCODE_ENTRY = 0
+from floss.string_decoder import decode_strings
 
 SIGNATURES_PATH_DEFAULT_STRING = "(embedded signatures)"
 EXTENSIONS_SHELLCODE_32 = ("sc32", "raw32")
 EXTENSIONS_SHELLCODE_64 = ("sc64", "raw64")
 
-logger = floss.logging.getLogger("floss")
+logger = floss.logging_.getLogger("floss")
+
+
+class StringType(str, Enum):
+    STATIC = "static"
+    DECODED = "decoded"
+    STACK = "stack"
+    TIGHT = "tight"
 
 
 class WorkspaceLoadError(ValueError):
     pass
-
-
-def set_vivisect_log_level(level):
-    logging.getLogger("vivisect").setLevel(level)
-    logging.getLogger("vivisect.base").setLevel(level)
-    logging.getLogger("vivisect.impemu").setLevel(level)
-    logging.getLogger("vtrace").setLevel(level)
-    logging.getLogger("envi").setLevel(level)
-    logging.getLogger("envi.codeflow").setLevel(level)
-
-
-def decode_strings(
-    vw,
-    functions: List[int],
-    min_length: int,
-    no_filter=False,
-    max_instruction_count=20000,
-    max_hits=1,
-    disable_progress=False,
-) -> List[DecodedString]:
-    """
-    FLOSS string decoding algorithm
-
-    arguments:
-        vw: the workspace
-        functions: addresses of the candidate decoding routines
-        min_length: minimun string length
-        max_instruction_count: max number of instructions to emulate per function
-        max_hits: max number of emulations per instruction
-    """
-    decoded_strings = []
-    function_index = viv_utils.InstructionFunctionIndex(vw)
-
-    pbar = tqdm.tqdm
-    if disable_progress:
-        # do not use tqdm to avoid unnecessary side effects when caller intends
-        # to disable progress completely
-        pbar = lambda s, *args, **kwargs: s
-
-    pb = pbar(functions, desc="decoding strings", unit=" functions")
-    for fva in pb:
-        for ctx in string_decoder.extract_decoding_contexts(vw, fva, max_hits):
-            for delta in string_decoder.emulate_decoding_routine(vw, function_index, fva, ctx, max_instruction_count):
-                for delta_bytes in string_decoder.extract_delta_bytes(delta, ctx.decoded_at_va, fva):
-                    for decoded_string in string_decoder.extract_strings(delta_bytes, min_length, no_filter):
-                        decoded_strings.append(decoded_string)
-    return decoded_strings
-
-
-def sanitize_string_for_printing(s: str) -> str:
-    """
-    Return sanitized string for printing.
-    :param s: input string
-    :return: sanitized string
-    """
-    sanitized_string = s.replace("\\\\", "\\")  # print single backslashes
-    sanitized_string = "".join(c for c in sanitized_string if c in string.printable)
-    return sanitized_string
-
-
-def sanitize_string_for_script(s: str) -> str:
-    """
-    Return sanitized string that is added to IDAPython script content.
-    :param s: input string
-    :return: sanitized string
-    """
-    sanitized_string = sanitize_string_for_printing(s)
-    sanitized_string = sanitized_string.replace("\\", "\\\\")
-    sanitized_string = sanitized_string.replace('"', '\\"')
-    return sanitized_string
 
 
 class ArgumentValueError(ValueError):
@@ -134,54 +68,99 @@ class ArgumentValueError(ValueError):
 class ArgumentParser(argparse.ArgumentParser):
     """
     argparse will call sys.exit upon parsing invalid arguments.
-    we don't want that, because we might be parsing args within test cases, etc.
+    we don't want that, because we might be parsing args within test cases, run as a module, etc.
     so, we override the behavior to raise a ArgumentValueError instead.
-
-    note: the help message will still be printed to the console.
 
     this strategy is originally described here: https://stackoverflow.com/a/16942165/87207
     """
 
     def error(self, message):
-        self.print_help(sys.stderr)
-        raise ArgumentValueError("%s: error: %s\n" % (self.prog, message))
+        self.print_usage(sys.stderr)
+        args = {"prog": self.prog, "message": message}
+        raise ArgumentValueError("%(prog)s: error: %(message)s" % args)
 
 
 def make_parser(argv):
-    desc = "The FLARE team's open-source tool to extract obfuscated strings from malware.\n  %(prog)s {:s} - https://github.com/mandiant/flare-floss/".format(
-        __version__
+    desc = (
+        "The FLARE team's open-source tool to extract obfuscated strings from malware.\n"
+        f"  %(prog)s {__version__} - https://github.com/mandiant/flare-floss/\n\n"
+        "FLOSS extracts all the following string types:\n"
+        ' 1. static strings:  "regular" ASCII and UTF-16LE strings\n'
+        " 2. decoded strings: strings decoded in a function\n"
+        " 3. stack strings:   strings constructed on the stack at run-time\n"
+        " 4. tight strings:   special form of stack strings, decoded on the stack\n"
     )
     epilog = textwrap.dedent(
         """
-        only displaying core arguments, run `floss --help -x` to see all supported arguments
+        only displaying core arguments, run `floss -H` to see all supported options
 
         examples:
           extract all strings from an executable
             floss suspicious.exe
 
-          extract all strings from 32-bit shellcode
-            floss -f sc32 shellcode.bin
+          do not extract static strings
+            floss --no static -- suspicious.exe
+
+          only extract stack and tight strings
+            floss --only stack tight -- suspicious.exe
         """
     )
-    epilog_expert = textwrap.dedent(
+    epilog_advanced = textwrap.dedent(
         """
         examples:
-          only show strings of minimun length 6
-            floss -n 6 suspicious.exe
+          extract all strings from 32-bit shellcode
+            floss -f sc32 shellcode.bin
 
-          only show stack strings
-            floss --no-static-strings --no-decoded-strings suspicious.exe
+          only decode strings from the specified functions
+            floss --functions 0x401000 0x401100 suspicious.exe
         """
     )
 
-    expert = "-x" in argv
+    show_all_options = "-H" in argv
 
     parser = ArgumentParser(
         description=desc,
-        epilog=epilog_expert if expert else epilog,
+        epilog=epilog_advanced if show_all_options else epilog,
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
+    parser.register("action", "extend", floss.utils.ExtendAction)
+    parser.add_argument("-H", action="help", help="show advanced options and exit")
 
+    parser.add_argument(
+        "-n",
+        "--minimum-length",
+        dest="min_length",
+        default=MIN_STRING_LENGTH,
+        help="minimum string length",
+    )
+
+    parser.add_argument(
+        "sample",
+        type=argparse.FileType("rb"),
+        help="path to sample to analyze",
+    )
+
+    analysis_group = parser.add_argument_group("analysis arguments")
+    analysis_group.add_argument(
+        "--no",
+        action="extend",
+        dest="disabled_types",
+        nargs="+",
+        choices=[t.value for t in StringType],
+        default=[],
+        help="do not extract specified string type(s)",
+    )
+    analysis_group.add_argument(
+        "--only",
+        action="extend",
+        dest="enabled_types",
+        nargs="+",
+        choices=[t.value for t in StringType],
+        default=[],
+        help="only extract specified string type(s)",
+    )
+
+    advanced_group = parser.add_argument_group("advanced arguments")
     formats = [
         ("auto", "(default) detect file type automatically"),
         ("pe", "Windows PE file"),
@@ -189,135 +168,79 @@ def make_parser(argv):
         ("sc64", "64-bit shellcode"),
     ]
     format_help = ", ".join(["%s: %s" % (f[0], f[1]) for f in formats])
-    parser.add_argument(
+    advanced_group.add_argument(
         "-f",
         "--format",
         choices=[f[0] for f in formats],
         default="auto",
-        help="select sample format, %s" % format_help,
+        help="select sample format, %s" % format_help if show_all_options else argparse.SUPPRESS,
     )
-
-    parser.add_argument(
-        "-n",
-        "--minimum-length",
-        dest="min_length",
-        default=DEFAULT_MIN_LENGTH,
-        help="minimum string length" if expert else argparse.SUPPRESS,
-    )
-
-    parser.add_argument(
-        "--signatures",
-        type=str,
-        default=SIGNATURES_PATH_DEFAULT_STRING,
-        help="path to .sig/.pat file or directory used to identify library functions, use embedded signatures by default"
-        if expert
-        else argparse.SUPPRESS,
-    )
-
-    parser.add_argument("-x", action="store_true", dest="x", help="enable eXpert arguments, see `floss --help -x`")
-    parser.add_argument(
-        "--version", action="version", version="%(prog)s {:s}".format(__version__), help=argparse.SUPPRESS
-    )
-
-    parser.add_argument(
-        "sample",
-        type=str,
-        help="path to sample to analyze",
-    )
-
-    # TODO move group to first position
-    analysis_group = parser.add_argument_group("analysis arguments")
-    analysis_group.add_argument(
+    advanced_group.add_argument(
         "--functions",
         type=lambda x: int(x, 0x10),
         default=None,
         nargs="+",
         help="only analyze the specified functions, hex-encoded like 0x401000, space-separate multiple functions"
-        if expert
+        if show_all_options
         else argparse.SUPPRESS,
     )
-
-    analysis_group.add_argument(
-        "--no-filter",
+    advanced_group.add_argument(
+        "--disable-progress",
         action="store_true",
-        help="do not filter deobfuscated strings (may result in many false positive strings)"
-        if expert
+        help="disable all progress bars" if show_all_options else argparse.SUPPRESS,
+    )
+    advanced_group.add_argument(
+        "--signatures",
+        type=str,
+        default=SIGNATURES_PATH_DEFAULT_STRING,
+        help="path to .sig/.pat file or directory used to identify library functions, use embedded signatures by default"
+        if show_all_options
         else argparse.SUPPRESS,
     )
-
-    analysis_group.add_argument(
-        "--max-instruction-count",
-        type=int,
-        default=DEFAULT_MAX_INSN_COUNT,
-        help="maximum number of instructions to emulate per function" if expert else argparse.SUPPRESS,
-    )
-
-    analysis_group.add_argument(
-        "--max-address-revisits",
-        type=int,
-        default=DEFAULT_MAX_ADDRESS_REVISITS,
-        help="maximum number of address revisits per function" if expert else argparse.SUPPRESS,
-    )
-
-    analysis_group.add_argument(
-        "--no-static-strings",
-        action="store_true",
-        default=False,
-        help="do not show static ASCII and UTF-16 strings" if expert else argparse.SUPPRESS,
-    )
-    analysis_group.add_argument(
-        "--no-decoded-strings",
-        action="store_true",
-        default=False,
-        help="do not show decoded strings" if expert else argparse.SUPPRESS,
-    )
-    analysis_group.add_argument(
-        "--no-stack-strings",
-        action="store_true",
-        default=False,
-        help="do not show stackstrings" if expert else argparse.SUPPRESS,
-    )
-    analysis_group.add_argument(
-        "--no-tight-strings",
-        action="store_true",
-        default=False,
-        help="do not show tightstrings" if expert else argparse.SUPPRESS,
+    advanced_group.add_argument(
+        "--version",
+        action="version",
+        version="%(prog)s {:s}".format(__version__),
+        help="show program's version number and exit" if show_all_options else argparse.SUPPRESS,
     )
 
     output_group = parser.add_argument_group("rendering arguments")
     output_group.add_argument("-j", "--json", action="store_true", help="emit JSON instead of text")
     output_group.add_argument(
-        "-v", "--verbose", action="store_true", help="enable verbose result document (no effect with --json)"
+        "-v",
+        "--verbose",
+        action="count",
+        default=floss.render.default.Verbosity.DEFAULT,
+        help="enable verbose results, e.g. including function offsets (does not affect JSON output)",
     )
     output_group.add_argument(
-        "-vv", "--vverbose", action="store_true", help="enable very verbose result document (no effect with --json)"
-    )
-    # TODO currently unused
-    output_group.add_argument(
-        "--color",
+        "-o",
+        "--outfile",
         type=str,
-        choices=("auto", "always", "never"),
-        default="auto",
-        help="enable ANSI color codes in results, default: only during interactive session",
+        help="write results to output file, instead of default STDOUT",
     )
 
     logging_group = parser.add_argument_group("logging arguments")
-
-    logging_group.add_argument("-d", "--debug", action="store_true", help="enable debugging output on STDERR")
-    logging_group.add_argument("-dd", "--trace", action="store_true", help="enable verbose debugging output on STDERR")
     logging_group.add_argument(
-        "-q", "--quiet", action="store_true", help="disable all status output except fatal errors"
+        "-d",
+        "--debug",
+        action="count",
+        default=DebugLevel.NONE,
+        help="enable debugging output on STDERR, specify multiple times to increase verbosity",
+    )
+    logging_group.add_argument(
+        "-q", "--quiet", action="store_true", help="disable all status output on STDOUT except fatal errors"
     )
 
     return parser
 
 
-def set_log_config(args):
-    if args.quiet:
+def set_log_config(debug, quiet):
+    if quiet:
         log_level = logging.WARNING
-    elif args.trace:
+    elif debug >= DebugLevel.TRACE:
         log_level = logging.TRACE
-    elif args.debug:
+    elif debug >= DebugLevel.DEFAULT:
         log_level = logging.DEBUG
     else:
         log_level = logging.INFO
@@ -325,31 +248,32 @@ def set_log_config(args):
     logging.basicConfig(level=log_level)
     logging.getLogger().setLevel(log_level)
 
-    # TODO enable and do more testing
-    # disable vivisect-related logging, it's verbose and not relevant for FLOSS users
-    set_vivisect_log_level(logging.CRITICAL)
+    if debug < DebugLevel.SUPERTRACE:
+        # these loggers are too verbose even for the TRACE level, enable via `-ddd`
+        logging.getLogger("floss.api_hooks").setLevel(logging.WARNING)
+        logging.getLogger("floss.function_argument_getter").setLevel(logging.WARNING)
+
+    # configure vivisect-related logging, it's verbose and not relevant for regular FLOSS users
+    # enable to do more vigorous testing
+    if debug < DebugLevel.TRACE:
+        set_vivisect_log_level(logging.CRITICAL)
+    else:
+        set_vivisect_log_level(logging.DEBUG)
+
+    # configure viv-utils logging
+    if debug == DebugLevel.DEFAULT:
+        logging.getLogger("Monitor").setLevel(logging.DEBUG)
+        logging.getLogger("EmulatorDriver").setLevel(logging.DEBUG)
+    elif debug <= DebugLevel.TRACE:
+        logging.getLogger("Monitor").setLevel(logging.ERROR)
+        logging.getLogger("EmulatorDriver").setLevel(logging.ERROR)
 
     # install the log message colorizer to the default handler.
     # because basicConfig is just above this,
     # handlers[0] is a StreamHandler to STDERR.
     #
     # calling this code from outside script main may do something unexpected.
-    logging.getLogger().handlers[0].setFormatter(floss.logging.ColorFormatter())
-
-
-def validate_sample_path(parser, args) -> str:
-    """
-    Return validated input file path or terminate program.
-    """
-    try_help_msg = "Try 'floss -h' for more information"
-
-    if not os.path.exists(args.sample):
-        parser.error("File '%s' does not exist\n%s" % (args.sample, try_help_msg))
-
-    if not os.path.isfile(args.sample):
-        parser.error("'%s' is not a file\n%s" % (args.sample, try_help_msg))
-
-    return args.sample
+    logging.getLogger().handlers[0].setFormatter(floss.logging_.ColorFormatter())
 
 
 def select_functions(vw, asked_functions: Optional[List[int]]) -> Set[int]:
@@ -383,17 +307,6 @@ def select_functions(vw, asked_functions: Optional[List[int]]) -> Set[int]:
     return asked_functions_
 
 
-def filter_unique_decoded(decoded_strings):
-    unique_values = set()
-    originals = []
-    for decoded in decoded_strings:
-        hashable = (decoded.string, decoded.decoded_at, decoded.decoding_routine)
-        if hashable not in unique_values:
-            unique_values.add(hashable)
-            originals.append(decoded)
-    return originals
-
-
 def is_workspace_file(sample_file_path):
     """
     Return if input file is a vivisect workspace, based on file extension
@@ -418,109 +331,6 @@ def is_supported_file_type(sample_file_path):
         return True
     else:
         return False
-
-
-def print_decoding_results(decoded_strings: List[DecodedString], quiet=False):
-    """
-    Print results of string decoding phase.
-
-    :param decoded_strings: list of decoded strings ([DecodedString])
-    :param quiet: print strings only, suppresses headers
-    """
-    logger.info("decoded %d strings" % len(decoded_strings))
-    fvas = set([i.decoding_routine for i in decoded_strings])
-    for fva in fvas:
-        grouped_strings = [ds for ds in decoded_strings if ds.decoding_routine == fva]
-        len_ds = len(grouped_strings)
-        if len_ds > 0:
-            logger.info("using decoding function at 0x%X (decoded %d strings):" % (fva, len_ds))
-            print_decoded_strings(grouped_strings, quiet=quiet)
-
-
-def print_decoded_strings(decoded_strings: List[DecodedString], quiet=False):
-    """
-    Print decoded strings.
-    :param decoded_strings: list of decoded strings ([DecodedString])
-    :param quiet: print strings only, suppresses headers
-    """
-    if quiet:
-        for ds in decoded_strings:
-            print(sanitize_string_for_printing(ds.string))
-    else:
-        ss = []
-        for ds in decoded_strings:
-            s = sanitize_string_for_printing(ds.string)
-            if ds.address_type == AddressType.STACK:
-                offset_string = "[STACK]"
-            elif ds.address_type == AddressType.HEAP:
-                offset_string = "[HEAP]"
-            else:
-                offset_string = hex(ds.address or 0)
-            ss.append((offset_string, hex(ds.decoded_at), s))
-
-        if len(ss) > 0:
-            print(tabulate.tabulate(ss, headers=["Offset", "Called At", "String"]))
-
-
-def get_file_as_mmap(path):
-    """
-    Returns an mmap object of the file
-    :param path: path of the file to map
-    """
-    with open(path, "rb") as f:
-        return mmap.mmap(f.fileno(), 0, access=mmap.ACCESS_READ)
-
-
-# TODO: move to floss.render
-def print_static_strings(results: ResultDocument):
-    """
-    Print static ASCII and UTF-16 strings from provided file.
-    """
-    ascii_strings = [s.string for s in results.strings.static_strings if s.encoding == StringEncoding.ASCII]
-    unicode_strings = [s.string for s in results.strings.static_strings if s.encoding == StringEncoding.UTF16LE]
-
-    if not ascii_strings:
-        print("static ASCII strings (%d): none" % (len(unicode_strings)))
-    else:
-        print("static ASCII strings (%d):" % (len(ascii_strings)))
-        for s in ascii_strings:
-            print(s)
-        print()
-
-    if not unicode_strings:
-        print("static UTF-16LE strings (%d): none" % (len(unicode_strings)))
-    else:
-        print("static UTF-16LE strings (%d):" % (len(unicode_strings)))
-        for s in unicode_strings:
-            print(s)
-
-
-def print_stack_strings(extracted_strings: Union[List[StackString], List[TightString]], quiet=False):
-    """
-    Print extracted stackstrings.
-    :param extracted_strings: list of stack strings ([StackString])
-    :param quiet: print strings only, suppresses headers
-    """
-    count = len(extracted_strings)
-
-    logger.info("FLOSS extracted %d stackstrings" % (count))
-
-    if quiet:
-        for ss in extracted_strings:
-            print("%s" % (ss.string))
-    elif count > 0:
-        print(
-            tabulate.tabulate(
-                [(hex(s.function), hex(s.frame_offset), s.string) for s in extracted_strings],
-                headers=["Function", "Frame Offset", "String"],
-            )
-        )
-
-
-def print_file_meta_info(vw, selected_functions: Set[int]):
-    logger.trace("analysis summary:")
-    for k, v in get_vivisect_meta_info(vw, selected_functions).items():
-        logger.trace("  %s: %s", k, v or "N/A")
 
 
 class Architecture(str, Enum):
@@ -622,21 +432,34 @@ def get_signatures(sigs_path):
     return paths
 
 
+def is_string_type_enabled(type_, disabled_types, enabled_types):
+    if disabled_types:
+        return type_ not in disabled_types
+    elif enabled_types:
+        return type_ in enabled_types
+    else:
+        return True
+
+
 def main(argv=None) -> int:
     """
     arguments:
-      argv: the command line arguments, including the executable name, like sys.argv
+      argv: the command line arguments
     """
-    if not argv:
-        argv = sys.argv
+    if argv is None:
+        argv = sys.argv[1:]
 
-    parser = make_parser(argv[1:])
+    parser = make_parser(argv)
     try:
-        args = parser.parse_args(args=argv[1:])
+        args = parser.parse_args(args=argv)
+        # manual check here, because add_mutually_exclusive_group() on argument_group("...") appears wrong
+        if args.enabled_types and args.disabled_types:
+            parser.error("--no and --only arguments are not allowed together")
     except ArgumentValueError as e:
+        print(e)
         return -1
 
-    set_log_config(args)
+    set_log_config(args.debug, args.quiet)
 
     # Since Python 3.8 cp65001 is an alias to utf_8, but not for Python < 3.8
     # TODO: remove this code when only supporting Python 3.8+
@@ -659,22 +482,32 @@ def main(argv=None) -> int:
 
         args.signatures = sigs_path
 
-    sample = validate_sample_path(parser, args)
+    # TODO pass buffer along instead of file path, also should work for stdin
+    sample = args.sample.name
+    args.sample.close()
+
+    if args.functions:
+        # when analyzing specified functions do not show static strings
+        args.disabled_types.append(StringType.STATIC)
 
     results = ResultDocument(
         metadata=Metadata(
             file_path=sample,
-            enable_stack_strings=not args.no_stack_strings,
-            enable_tight_strings=not args.no_tight_strings,
-            enable_decoded_strings=not args.no_decoded_strings,
-            enable_static_strings=not args.no_static_strings,
+            enable_static_strings=is_string_type_enabled(StringType.STATIC, args.disabled_types, args.enabled_types),
+            enable_stack_strings=is_string_type_enabled(StringType.STACK, args.disabled_types, args.enabled_types),
+            enable_decoded_strings=is_string_type_enabled(StringType.DECODED, args.disabled_types, args.enabled_types),
+            enable_tight_strings=is_string_type_enabled(StringType.TIGHT, args.disabled_types, args.enabled_types),
         )
     )
 
-    # 1. static strings, because its fast
-    # 2. decoded strings
-    # 3. stack strings  # TODO move to 2. since it's also fast
-    # 4. tight strings
+    time0 = time()
+    interim = time0
+
+    # in order of expected run time, fast to slow
+    # 1. static strings
+    # 2. stack strings
+    # 3. tight strings
+    # 4. decoded strings
 
     if results.metadata.enable_static_strings:
         logger.info("extracting static strings...")
@@ -683,13 +516,11 @@ def main(argv=None) -> int:
 
         with open(sample, "rb") as f:
             with contextlib.closing(mmap.mmap(f.fileno(), 0, access=mmap.ACCESS_READ)) as buf:
-                static_ascii_strings = list(strings.extract_ascii_strings(buf, args.min_length))
-                static_unicode_strings = list(strings.extract_unicode_strings(buf, args.min_length))
+                static_strings = list(extract_ascii_unicode_strings(buf, args.min_length))
 
-        results.strings.static_strings = static_ascii_strings + static_unicode_strings
-
-        if not args.json:
-            print_static_strings(results)
+        results.strings.static_strings = static_strings
+        results.metadata.runtime.static_strings = get_runtime_diff(interim)
+        interim = time()
 
     if (
         results.metadata.enable_decoded_strings
@@ -704,8 +535,15 @@ def main(argv=None) -> int:
 
         should_save_workspace = os.environ.get("FLOSS_SAVE_WORKSPACE") not in ("0", "no", "NO", "n", None)
         try:
-            with halo.Halo(text="analyzing program", spinner="simpleDots", stream=sys.stderr, enabled=not args.quiet):
+            with halo.Halo(
+                text="analyzing program",
+                spinner="simpleDots",
+                stream=sys.stderr,
+                enabled=not (args.quiet or args.disable_progress),
+            ):
                 vw = load_vw(sample, args.format, sigpaths, should_save_workspace)
+                results.metadata.runtime.vivisect = get_runtime_diff(interim)
+                interim = time()
         except WorkspaceLoadError as e:
             logger.error("failed to analyze sample: %s", e)
             return -1
@@ -720,74 +558,89 @@ def main(argv=None) -> int:
             logger.error(e.args[0])
             return -1
 
-        logger.trace("analysis summary:")
-        for k, v in get_vivisect_meta_info(vw, selected_functions).items():
-            logger.trace("  %s: %s", k, v or "N/A")
-
-        time0 = time()
-
-        logger.info("identifying decoding function features...")
         decoding_function_features, meta_lib_funcs = find_decoding_function_features(
-            vw, selected_functions, disable_progress=args.quiet
+            vw, selected_functions, disable_progress=args.quiet or args.disable_progress
         )
         results.metadata.analysis.update(meta_lib_funcs)
+        results.metadata.runtime.find_features = get_runtime_diff(interim)
+        interim = time()
 
-        if results.metadata.enable_decoded_strings:
-            top_functions = get_top_functions(decoding_function_features, 20)
-            # TODO also emulate tightfuncs that have a tight loop and are short < 5 BBs
-
-            if len(top_functions) == 0:
-                logger.info("no candidate decoding functions found.")
-            else:
-                logger.info("identified %d candidate decoding functions", len(top_functions))
-                for fva, function_data in top_functions:
-                    logger.debug("  - 0x%x: %.3f", fva, function_data["score"])
-
-            logger.info("decoding strings...")
-            results.strings.decoded_strings = decode_strings(
-                vw,
-                list(map(lambda p: p[0], top_functions)),
-                args.min_length,
-                False,
-                args.max_instruction_count,
-                args.max_address_revisits + 1,
-                disable_progress=args.quiet,
-            )
-            # TODO: The de-duplication process isn't perfect as it is done here and in print_decoding_results and
-            #       all of them on non-sanitized strings.
-            results.strings.decoded_strings = filter_unique_decoded(results.strings.decoded_strings)
-            if not args.json:
-                print_decoding_results(results.strings.decoded_strings, quiet=args.quiet)
+        logger.trace("analysis summary:")
+        for k, v in get_vivisect_meta_info(vw, selected_functions, decoding_function_features).items():
+            logger.trace("  %s: %s", k, v or "N/A")
 
         if results.metadata.enable_stack_strings:
-            # don't run this on functions with tight loops as this will likely result in FPs
-            # and should be caught by the tightstrings extraction below
-            no_tightloop_functions = get_functions_without_tightloops(decoding_function_features)
+            if results.metadata.enable_tight_strings:
+                # don't run this on functions with tight loops as this will likely result in FPs
+                # and should be caught by the tightstrings extraction below
+                selected_functions = get_functions_without_tightloops(decoding_function_features)
 
-            logger.info("extracting stackstrings...")
-            results.strings.stack_strings = list(
-                stackstrings.extract_stackstrings(vw, no_tightloop_functions, args.min_length, quiet=args.quiet)
+            results.strings.stack_strings = extract_stackstrings(
+                vw,
+                selected_functions,
+                args.min_length,
+                verbosity=args.verbose,
+                disable_progress=args.quiet or args.disable_progress,
             )
-
-            # remove duplicate entries
-            results.strings.stack_strings = list(set(results.strings.stack_strings))
-            if not args.json:
-                print_stack_strings(results.strings.stack_strings, quiet=args.quiet)
+            results.metadata.runtime.stack_strings = get_runtime_diff(interim)
+            interim = time()
 
         if results.metadata.enable_tight_strings:
-            logger.info("extracting tightstrings...")
             tightloop_functions = get_functions_with_tightloops(decoding_function_features)
             # TODO if there are many tight loop functions, emit that the program likely uses tightstrings? see #400
-            results.strings.tight_strings = list(extract_tightstrings(vw, tightloop_functions, quiet=args.quiet))
+            results.strings.tight_strings = list(
+                extract_tightstrings(
+                    vw,
+                    tightloop_functions,
+                    min_length=args.min_length,
+                    verbosity=args.verbose,
+                    disable_progress=args.quiet or args.disable_progress,
+                )
+            )
+            results.metadata.runtime.tight_strings = get_runtime_diff(interim)
+            interim = time()
 
-            if not args.json:
-                print_stack_strings(results.strings.tight_strings, quiet=args.quiet)
+        if results.metadata.enable_decoded_strings:
+            # TODO select more based on score rather than absolute count?!
+            top_functions = get_top_functions(decoding_function_features, 20)
 
-        time1 = time()
-        logger.info("finished execution after %f seconds", (time1 - time0))
+            fvas_to_emulate = get_function_fvas(top_functions)
+            fvas_tight_functions = get_tight_function_fvas(
+                decoding_function_features
+            )  # TODO exclude from stackstrings?!
+            fvas_to_emulate = append_unique(fvas_to_emulate, fvas_tight_functions)
 
-        if args.json:
-            print(floss.render.json.render(results))
+            if len(fvas_to_emulate) == 0:
+                logger.info("no candidate decoding functions found.")
+            else:
+                logger.debug("identified %d candidate decoding functions", len(fvas_to_emulate))
+                for fva in fvas_to_emulate:
+                    logger.debug("  - 0x%x: %.3f", fva, decoding_function_features[fva]["score"])
+
+            # TODO filter out strings decoded in library function or function only called by library function(s)
+            results.strings.decoded_strings = decode_strings(
+                vw,
+                fvas_to_emulate,
+                args.min_length,
+                verbosity=args.verbose,
+                disable_progress=args.quiet or args.disable_progress,
+            )
+            results.metadata.runtime.decoded_strings = get_runtime_diff(interim)
+
+    results.metadata.runtime.total = get_runtime_diff(time0)
+    logger.info("finished execution after %.2f seconds", results.metadata.runtime.total)
+
+    if args.json:
+        r = floss.render.json.render(results)
+    else:
+        r = floss.render.default.render(results, args.verbose, args.quiet)
+
+    if args.outfile:
+        logger.info("writing results to %s", args.outfile)
+        with open(args.outfile, "wb") as f:
+            f.write(r.encode("utf-8"))
+    else:
+        print(r)
 
     return 0
 
