@@ -1,65 +1,68 @@
-from itertools import chain
+from typing import Set, List, Tuple, Iterator, Optional
 
 import tqdm
 import viv_utils
 import tqdm.contrib.logging
+import viv_utils.emulator_drivers
 
 import floss.utils
 import floss.features.features
-from floss import stackstrings
-from floss.results import TightString, StaticString
-from floss.strings import extract_ascii_strings, extract_unicode_strings
-from floss.stackstrings import StackstringContextMonitor
+from floss.const import TS_MAX_INSN_COUNT
+from floss.utils import extract_strings
+from floss.results import TightString
+from floss.stackstrings import CallContext, StackstringContextMonitor
+from floss.render.default import Verbosity
 
-logger = floss.logging.getLogger(__name__)
+logger = floss.logging_.getLogger(__name__)
 
 
 class TightstringContextMonitor(StackstringContextMonitor):
-    def __init__(self, vw, sp, tloops):
+    def __init__(self, vw, sp, min_length):
         super(TightstringContextMonitor, self).__init__(vw, sp, [])
-        self.tloop_startvas = [t.startva for t in tloops]
-        self.tloop_endvas = [t.endva for t in tloops]
-        # store FP stackstrings before tightstring loop executes
-        self.pre_ctx_strings = set()
-        logger.trace(" stavas: %s", ", ".join(map(hex, self.tloop_startvas)))
-        logger.trace(" endvas: %s", ", ".join(map(hex, self.tloop_endvas)))
+        self.min_length = min_length
 
+    # override base class
     def apicall(self, emu, op, pc, api, argv):
         pass
 
-    def prehook(self, emu, op, startpc):
-        if startpc in self.tloop_startvas:
-            try:
-                stack_buf = self.get_call_context(emu, op).stack_memory
-            except ValueError as e:
-                logger.debug(str(e))
-                return
+    def get_pre_ctx_strings(self, emu) -> Set[str]:
+        try:
+            stack_buf = self.get_call_context(emu, emu.getProgramCounter()).stack_memory
+            return set(map(lambda s: s.string, extract_strings(stack_buf, self.min_length)))
+        except ValueError as e:
+            logger.debug("%s", e)
+        return set()
 
-            self.pre_ctx_strings.update(map(lambda s: s.string, floss.strings.extract_ascii_strings(stack_buf)))
-            self.pre_ctx_strings.update(map(lambda s: s.string, floss.strings.extract_unicode_strings(stack_buf)))
-
-            # only save one context per tightloop
-            self.tloop_startvas.remove(startpc)
-
-    def posthook(self, emu, op, endpc):
-        if endpc in self.tloop_endvas:
-            logger.trace("extracting context at endpc: 0x%x", endpc)
-            self.extract_context(emu, op)
-
-            # only extract once at tightloop end
-            self.tloop_endvas.remove(endpc)
+    def get_context(self, emu, va, pre_ctx_strings: Optional[Set[str]]) -> Iterator[CallContext]:
+        try:
+            yield self.get_call_context(emu, va, pre_ctx_strings)
+        except ValueError as e:
+            logger.debug("%s", e)
 
 
-def extract_tightstring_contexts(vw, fva, tloops):
+def extract_tightstring_contexts(vw, fva, min_length, tloops) -> Iterator[CallContext]:
     emu = floss.utils.make_emulator(vw)
-    monitor = TightstringContextMonitor(vw, emu.getStackCounter(), tloops)
-    driver = viv_utils.emulator_drivers.FunctionRunnerEmulatorDriver(emu)
-    driver.add_monitor(monitor)
-    driver.runFunction(fva, maxhit=0x100, maxrep=0x100, func_only=True)
-    return monitor.ctxs, monitor.pre_ctx_strings
+    monitor = TightstringContextMonitor(vw, emu.getStackCounter(), min_length)
+    driver_all_paths = viv_utils.emulator_drivers.FunctionRunnerEmulatorDriver(emu)
+    driver_all_paths.add_monitor(monitor)
+    driver = viv_utils.emulator_drivers.DebuggerEmulatorDriver(emu)
+
+    for t in tloops:
+        # find and emulate single path to start of tight loop
+        driver_all_paths.runFunction(fva, t.startva, maxhit=1, maxrep=0x100, func_only=True)
+        # find existing (FP) stackstrings before tightstring loop executes
+        pre_ctx_strings = monitor.get_pre_ctx_strings(emu)
+        try:
+            # emulate tight loop
+            driver.runToVa(t.endva, max_instruction_count=TS_MAX_INSN_COUNT)
+        except Exception as e:
+            logger.debug("error emulating tight loop starting at 0x%x in function 0x%x: %s", t.startva, fva, e)
+        yield from monitor.get_context(emu, t.startva, pre_ctx_strings)
 
 
-def extract_tightstrings(vw, tightloop_functions, quiet=False):
+def extract_tightstrings(
+    vw, tightloop_functions, min_length, verbosity=Verbosity.DEFAULT, disable_progress=False
+) -> List[TightString]:
     """
     Extracts tightstrings from functions that contain tight loops.
     Tightstrings are a special form of stackstrings. Their bytes are loaded on the stack and then modified in a
@@ -69,49 +72,33 @@ def extract_tightstrings(vw, tightloop_functions, quiet=False):
 
     :param vw: The vivisect workspace
     :param tightloop_functions: functions containing tight loops
-    :param quiet: do NOT show progress bar
-    :rtype: Generator[StackString]
+    :param min_length: minimum string length
+    :param verbosity: verbosity level
+    :param disable_progress: do NOT show progress bar
     """
-    # TODO add test sample(s) and tests
-    pbar = tqdm.tqdm
-    if quiet:
-        # do not use tqdm to avoid unnecessary side effects when caller intends
-        # to disable progress completely
-        pbar = lambda s, *args, **kwargs: s
+    logger.info("extracting tightstrings from %d functions...", len(tightloop_functions))
 
-    pb = pbar(tightloop_functions.items(), desc="extracting tightstrings", unit=" functions")
+    tight_strings = list()
+    # TODO add test sample(s) and tests
+    pb = floss.utils.get_progress_bar(
+        tightloop_functions.items(), disable_progress, desc="extracting tightstrings", unit=" functions"
+    )
     with tqdm.contrib.logging.logging_redirect_tqdm(), floss.utils.redirecting_print_to_tqdm():
         for fva, tloops in pb:
-            fva_s = f"0x{fva:x}"
-            pb.set_description(f"extracting tightstrings from {fva_s}")
-            with floss.utils.timing(fva_s):
-                logger.debug("extracting tightstrings from function: 0x%x", fva)
-                ctxs, pre_ctx_strings = extract_tightstring_contexts(vw, fva, tloops)
-                logger.trace("pre_ctx strings: %s", pre_ctx_strings)
-                for ctx in ctxs:
+            with floss.utils.timing(f"0x{fva:x}"):
+                logger.debug("extracting tightstrings from function 0x%x", fva)
+                if isinstance(pb, tqdm.tqdm):
+                    pb.set_description(f"extracting tightstrings from function 0x{fva:x}")
+
+                ctxs = extract_tightstring_contexts(vw, fva, min_length, tloops)
+                for n, ctx in enumerate(ctxs, 1):
                     logger.trace(
                         "extracting tightstring at checkpoint: 0x%x stacksize: 0x%x", ctx.pc, ctx.init_sp - ctx.sp
                     )
-                    for s in chain(
-                        floss.strings.extract_ascii_strings(ctx.stack_memory),
-                        floss.strings.extract_unicode_strings(ctx.stack_memory),
-                    ):
-
-                        # TODO sanitize/cleanup string?
-                        #  pVA and other prefixes
-                        # if floss.utils.is_fp_string(s.string):
-                        #     continue
-                        # stripped_string = floss.utils.strip_string(s.string)
-
-                        # TODO remove dups, e.g. in e5f5ad
-
-                        if s.string not in pre_ctx_strings:
-                            frame_offset = (ctx.init_sp - ctx.sp) - s.offset - stackstrings.getPointerSize(vw)
-                            ts = TightString(
-                                fva, s.string, s.encoding, ctx.pc, ctx.sp, ctx.init_sp, s.offset, frame_offset
-                            )
-                            # TODO option/format to log quiet and regular, this is verbose output here currently
-                            logger.info(
-                                "%s [%s] in 0x%x at frame offset 0x%x", ts.string, ts.encoding, fva, ts.frame_offset
-                            )
-                            yield ts
+                    logger.trace("pre_ctx strings: %s", ctx.pre_ctx_strings)
+                    for s in extract_strings(ctx.stack_memory, min_length, exclude=ctx.pre_ctx_strings):
+                        frame_offset = (ctx.init_sp - ctx.sp) - s.offset - floss.utils.getPointerSize(vw)
+                        ts = TightString(fva, s.string, s.encoding, ctx.pc, ctx.sp, ctx.init_sp, s.offset, frame_offset)
+                        floss.results.log_result(ts, verbosity)
+                        tight_strings.append(ts)
+    return tight_strings
