@@ -3,9 +3,10 @@ import networkx
 import vivisect
 import viv_utils
 from networkx import strongly_connected_components
-from envi.archs.i386.opconst import INS_MOV, INS_ROL, INS_ROR, INS_SHL, INS_SHR, INS_XOR
+from envi.archs.i386.opconst import INS_MOV, INS_ROL, INS_ROR, INS_SHL, INS_SHR, INS_XOR, INS_CALL
 
-import floss.logging
+import floss.logging_
+from floss.const import TS_TIGHT_FUNCTION_MAX_BLOCKS
 from floss.features.features import (
     Mov,
     Loop,
@@ -14,6 +15,8 @@ from floss.features.features import (
     CallsTo,
     NzxorLoop,
     TightLoop,
+    BlockCount,
+    TightFunction,
     KindaTightLoop,
     NzxorTightLoop,
 )
@@ -24,7 +27,7 @@ SECURITY_COOKIE_BYTES_DELTA = 0x40
 
 SHIFT_ROTATE_INS = (INS_SHL, INS_SHR, INS_ROL, INS_ROR)
 
-logger = floss.logging.getLogger(__name__)
+logger = floss.logging_.getLogger(__name__)
 
 
 def extract_insn_nzxor(f, bb, insn):
@@ -118,7 +121,11 @@ def extract_function_kinda_tight_loop(f):
     """
     cfg = viv_utils.CFG(f)
 
-    root_bb_va = cfg.get_root_basic_block()
+    try:
+        root_bb_va = cfg.get_root_basic_block().va
+    except KeyError:
+        # likely wrongly identified or analyzed function
+        return
     leaf_bb_vas = {bb.va for bb in cfg.get_leaf_basic_blocks()}
 
     for bb in f.basic_blocks:
@@ -163,10 +170,12 @@ def extract_function_kinda_tight_loop(f):
         # the BB that branches back to `bb`, either [a] or [c]
         # or None if a tight loop is not found.
         loop_bb = None
+        is_very_tight = False
 
         # find very tight loops: [a]->[a]
         for suc in succs:
             if suc.va == bb.va:
+                is_very_tight = True
                 loop_bb = bb
 
         # find semi tight loops: [a]->[c]->[a]
@@ -194,28 +203,42 @@ def extract_function_kinda_tight_loop(f):
         if not next_bb:
             continue
 
-        # Blaine's algorithm gets the block before the loop here
-        # additionally, he prunes the identified loops before processing further
-        # TODO prune loops that do not write memory
+        # ignore tight loops that call other functions
+        if contains_call(bb):
+            continue
 
-        yield KindaTightLoop(bb.va, next_bb.va)
+        if not writes_memory(bb):
+            continue
+
+        if is_very_tight:
+            yield TightLoop(bb.va, next_bb.va)
+        else:
+            yield KindaTightLoop(bb.va, next_bb.va)
 
 
-def extract_bb_tight_loop(f, bb):
-    """check basic block for tight loop indicators"""
-    if _bb_has_tight_loop(f, bb):
-        yield TightLoop(bb.va, bb.va + bb.size)
+def contains_call(bb):
+    for insn in bb.instructions:
+        if insn.opcode == INS_CALL:
+            return True
+    return False
 
 
-def _bb_has_tight_loop(f, bb):
-    """
-    parse tight loops, true if last instruction in basic block branches to bb start
-    """
-    if len(bb.instructions) > 0:
-        for bva, bflags in bb.instructions[-1].getBranches():
-            if bflags & envi.BR_COND:
-                if bva == bb.va:
-                    return True
+def writes_memory(bb):
+    for insn in bb.instructions:
+        # don't handle len(ops) == 0 for `rep movsb` or other unexpected instructions
+        if len(insn.opers) < 1:
+            continue
+
+        # these also cover amd64
+        if isinstance(
+            insn.opers[0],
+            (
+                envi.archs.i386.disasm.i386RegMemOper,
+                envi.archs.i386.disasm.i386ImmMemOper,
+                envi.archs.i386.disasm.i386SibOper,
+            ),
+        ):
+            return True
     return False
 
 
@@ -229,6 +252,17 @@ def abstract_nzxor_tightloop(features):
 def abstract_nzxor_loop(features):
     if any(isinstance(f, Nzxor) for f in features) and any(isinstance(f, Loop) for f in features):
         yield NzxorLoop()
+
+
+def abstract_tightfunction(features):
+    """
+    (Kinda) TightLoop and only a few basic blocks
+    """
+    if any(filter(lambda f: isinstance(f, (TightLoop, KindaTightLoop)), features)):
+        for block_count in filter(lambda f: isinstance(f, BlockCount), features):
+            if block_count.value < TS_TIGHT_FUNCTION_MAX_BLOCKS:
+                yield TightFunction()
+                return
 
 
 def extract_function_loop(f):
@@ -273,7 +307,8 @@ def extract_function_features(f):
             yield feature
 
 
-BASIC_BLOCK_HANDLERS = (extract_bb_tight_loop,)
+# currently none, but this can change
+BASIC_BLOCK_HANDLERS = ()
 
 
 def extract_basic_block_features(f, bb):
@@ -298,6 +333,7 @@ def extract_insn_features(f, bb, insn):
 ABSTRACTION_HANDLERS = (
     abstract_nzxor_loop,
     abstract_nzxor_tightloop,
+    abstract_tightfunction,
 )
 
 
