@@ -8,14 +8,14 @@ import viv_utils
 import viv_utils.emulator_drivers
 
 import floss.utils
-import floss.logging
+import floss.logging_
 import floss.api_hooks
 
 # TODO get return address from emu_snap
 FunctionContext = namedtuple("FunctionContext", ["emu_snap", "return_address", "decoded_at_va"])
 
 
-logger = floss.logging.getLogger(__name__)
+logger = floss.logging_.getLogger(__name__)
 
 
 class CallMonitor(viv_utils.emulator_drivers.Monitor):
@@ -111,6 +111,7 @@ class FunctionArgumentGetter(viv_utils.LoggingObject):
         monitor = CallMonitor(self.vivisect_workspace, target_fva)
         with installed_monitor(self.driver, monitor):
             with floss.api_hooks.defaultHooks(self.driver):
+                # TODO maxhit == 1 makes most sense for getting all simple contexts (no loops etc. in generation)
                 self.driver.runFunction(self.index[fva], maxhit=max_hits, maxrep=0x1000, func_only=True)
         contexts = monitor.get_contexts()
 
