@@ -1,10 +1,13 @@
 import string
 
 
-def sanitize_string_for_printing(s: str) -> str:
+def sanitize(s: str) -> str:
     """
     Return sanitized string for printing to cli.
     """
-    sanitized_string = s.replace("\\\\", "\\")  # print single backslashes
-    sanitized_string = "".join(c for c in sanitized_string if c in string.printable)
-    return sanitized_string
+    s = s.replace("\n", "\\n")
+    s = s.replace("\r", "\\r")
+    s = s.replace("\t", "\\t")
+    s = s.replace("\\\\", "\\")  # print single backslashes
+    s = "".join(c for c in s if c in string.printable)
+    return s
