@@ -1,2 +1,2 @@
 # Copyright (C) 2017 Mandiant, Inc. All Rights Reserved.
-__version__ = "2.0.0-rc1"
+__version__ = "2.0.0"
