@@ -1,4 +1,6 @@
-#!/usr/bin/env python
+# TODO currently not maintained or tested
+
+#!/usr/bin/env python3
 """
 Run FLOSS to automatically extract obfuscated strings and apply them to the
 currently loaded module in IDA Pro.
@@ -16,6 +18,7 @@ import floss
 import floss.main
 import floss.identify
 import floss.stackstrings
+import floss.string_decoder
 import floss.decoding_manager
 from floss.results import AddressType
 
@@ -166,13 +169,17 @@ def main(argv=None):
     )
 
     logger.info("decoding strings...")
-    decoded_strings = floss.main.decode_strings(vw, decoding_functions_candidates, MIN_LENGTH, no_filter=True)
+    decoded_strings = floss.string_decoder.decode_strings(
+        vw, floss.identify.get_function_fvas(decoding_functions_candidates), MIN_LENGTH, disable_progress=True
+    )
     logger.info("decoded %d strings", len(decoded_strings))
 
     logger.info("extracting stackstrings...")
     stack_strings = floss.stackstrings.extract_stackstrings(vw, selected_functions, MIN_LENGTH, no_filter=True)
     stack_strings = set(stack_strings)
     logger.info("decoded %d stack strings", len(stack_strings))
+
+    # TODO tight strings
 
     apply_decoded_strings(decoded_strings)
 
